CREATE/post request created and tested successfully

+++ data/migrations/20191105152544_build-cars-table.js
@@ -6,7 +6,7 @@ exports.up = function(knex) {
     table.string('vin', 17).unique().notNullable();
     table.text('make', 128).notNullable();
     table.text('model', 128).notNullable();
-    table.decimal('mileage').notNullable();
+    table.integer('mileage').notNullable();
     table.text('transmissionType', 128);
     table.text('title', 128);
   })
